+++ api/src/identity-access-management/application/organization-learner-account-recovery/organization-learner-account-recovery.controller.js
@@ -1,0 +1,22 @@
+import { usecases } from '../../domain/usecases/index.js';
+import * as studentInformationForAccountRecoverySerializer from '../../infrastructure/serializers/jsonapi/student-information-for-account-recovery.serializer.js';
+
+async function checkScoAccountRecovery(request, h, dependencies = { studentInformationForAccountRecoverySerializer }) {
+  const studentInformation = await dependencies.studentInformationForAccountRecoverySerializer.deserialize(
+    request.payload,
+  );
+
+  const studentInformationForAccountRecovery = await usecases.checkScoAccountRecovery({
+    studentInformation,
+  });
+
+  return h.response(
+    dependencies.studentInformationForAccountRecoverySerializer.serialize(studentInformationForAccountRecovery),
+  );
+}
+
+const scoOrganizationLearnerController = {
+  checkScoAccountRecovery,
+};
+
+export { scoOrganizationLearnerController };
