+++ api/src/identity-access-management/application/organization-learner-account-recovery/organization-learner-account-recovery.route.js
@@ -1,0 +1,34 @@
+import JoiDate from '@joi/date';
+import BaseJoi from 'joi';
+const Joi = BaseJoi.extend(JoiDate);
+
+import { studentIdentifierType } from '../../../shared/domain/types/identifiers-type.js';
+import { scoOrganizationLearnerController } from './organization-learner-account-recovery.controller.js';
+
+export const organizationLearnerAccountRecoveryRoutes = [
+  {
+    method: 'POST',
+    path: '/api/sco-organization-learners/account-recovery',
+    config: {
+      auth: false,
+      handler: scoOrganizationLearnerController.checkScoAccountRecovery,
+      validate: {
+        payload: Joi.object({
+          data: {
+            attributes: {
+              'first-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+              'last-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+              'ine-ina': studentIdentifierType,
+              birthdate: Joi.date().format('YYYY-MM-DD').required(),
+            },
+          },
+        }).options({ allowUnknown: true }),
+      },
+      notes: [
+        "- Recherche d'un ancien élève par son ine/ina, prénom, nom, date de naissance \n" +
+          '- On renvoie les informations permettant de récupérer son compte Pix.',
+      ],
+      tags: ['api', 'sco-organization-learners'],
+    },
+  },
+];
