+++ api/src/identity-access-management/application/routes.js
@@ -3,6 +3,7 @@ import { anonymizationAdminRoutes } from './anonymization/anonymization.admin.ro
 import { ltiRoutes } from './lti/lti.route.js';
 import { oidcProviderAdminRoutes } from './oidc-provider/oidc-provider.admin.route.js';
 import { oidcProviderRoutes } from './oidc-provider/oidc-provider.route.js';
+import { organizationLearnerAccountRecoveryRoutes } from './organization-learner-account-recovery/organization-learner-account-recovery.route.js';
 import { passwordRoutes } from './password/password.route.js';
 import { samlRoutes } from './saml/saml.route.js';
 import { tokenRoutes } from './token/token.route.js';
@@ -15,6 +16,7 @@ const allRoutes = [
   ...ltiRoutes,
   ...oidcProviderAdminRoutes,
   ...oidcProviderRoutes,
+  ...organizationLearnerAccountRecoveryRoutes,
   ...passwordRoutes,
   ...samlRoutes,
   ...tokenRoutes,
