+++ api/src/identity-access-management/domain/usecases/check-sco-account-recovery.usecase.js
@@ -1,4 +1,4 @@
-import { StudentInformationForAccountRecovery } from '../../../src/shared/domain/read-models/StudentInformationForAccountRecovery.js';
+import { StudentInformationForAccountRecovery } from '../../../shared/domain/read-models/StudentInformationForAccountRecovery.js';
 
 const checkScoAccountRecovery = async function ({
   studentInformation,
