+++ api/tests/identity-access-management/unit/domain/usecases/check-sco-account-recovery_test.js
@@ -1,6 +1,6 @@
-import { checkScoAccountRecovery } from '../../../../lib/domain/usecases/check-sco-account-recovery.js';
-import { StudentInformationForAccountRecovery } from '../../../../src/shared/domain/read-models/StudentInformationForAccountRecovery.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { checkScoAccountRecovery } from '../../../../../src/identity-access-management/domain/usecases/check-sco-account-recovery.usecase.js';
+import { StudentInformationForAccountRecovery } from '../../../../../src/shared/domain/read-models/StudentInformationForAccountRecovery.js';
+import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | UseCase | check-sco-account-recovery', function () {
   let organizationLearnerRepository;
